separate webhook and runner components
This finishes off a point in the TODO which makes it easier to reason
about the difference between the webhook parsing and the command
running. This is useful because it means the command runner, which is
the most significant reason this project exists, can run distinct of
webhooks, and without caring about what data the webhook provides.